refactore UI and design : product Features

+++ lib/features/product/data/model/product_model.dart
@@ -28,8 +28,8 @@ class ProductModel extends ProductEntities {
       eId: data['e_id'],
       images: data['images'],
       type: data['type'],
-      purchasePrice: data['purchase_price'],
-      sellingPrice: data['selling_price'],
+      purchasePrice: data['purchase_price'].toDouble(),
+      sellingPrice: data['selling_price'].toDouble(),
       futureProduct: data["future_product"],
     );
   }
